+++ apps/meteor/app/push-notifications/server/methods/saveNotificationSettings.ts
@@ -1,10 +1,10 @@
-import { Meteor } from 'meteor/meteor';
-import { check } from 'meteor/check';
-import type { ServerMethods } from '@rocket.chat/ui-contexts';
 import type { ISubscription } from '@rocket.chat/core-typings';
 import { Subscriptions } from '@rocket.chat/models';
+import type { ServerMethods } from '@rocket.chat/ui-contexts';
+import { check } from 'meteor/check';
+import { Meteor } from 'meteor/meteor';
 
-import { getUserNotificationPreference } from '../../../utils/server';
+import { getUserNotificationPreference } from '../../../utils/server/getUserNotificationPreference';
 
 const saveAudioNotificationValue = (subId: ISubscription['_id'], value: string) =>
 	value === 'default' ? Subscriptions.clearAudioNotificationValueById(subId) : Subscriptions.updateAudioNotificationValueById(subId, value);