+++ apps/meteor/ee/app/livechat-enterprise/server/api/lib/tags.ts
@@ -1,6 +1,6 @@
-import { escapeRegExp } from '@rocket.chat/string-helpers';
-import { LivechatTag } from '@rocket.chat/models';
 import type { ILivechatTag } from '@rocket.chat/core-typings';
+import { LivechatTag } from '@rocket.chat/models';
+import { escapeRegExp } from '@rocket.chat/string-helpers';
 import type { Filter, FindOptions } from 'mongodb';
 
 import { hasPermissionAsync } from '../../../../../../app/authorization/server/functions/hasPermission';