+++ apps/meteor/ee/definition/rest/v1/sessions/index.ts
@@ -1,4 +1,0 @@
-import './sessions';
-
-export * from './SessionsProps';
-export * from './SessionsPaginateProps';