+++ packages/model-typings/src/models/IModerationReportsModel.ts
@@ -1,4 +1,4 @@
-import type { IModerationReport, IMessage, IModerationAudit } from '@rocket.chat/core-typings';
+import type { IModerationReport, IMessage, IModerationAudit, MessageReport } from '@rocket.chat/core-typings';
 import type { AggregationCursor, Document, FindCursor, FindOptions, UpdateResult } from 'mongodb';
 
 import type { FindPaginated, IBaseModel } from './IBaseModel';
@@ -30,7 +30,7 @@ export interface IModerationReportsModel extends IBaseModel<IModerationReport> {
 		pagination: PaginationParams<IModerationReport>,
 	): AggregationCursor<IModerationAudit>;
 
-	countReportsInRange(latest: Date, oldest: Date, selector: string): Promise<number>;
+	countMessageReportsInRange(latest: Date, oldest: Date, selector: string): Promise<number>;
 
 	findReportsByMessageId(
 		messageId: IMessage['_id'],
@@ -44,9 +44,14 @@ export interface IModerationReportsModel extends IBaseModel<IModerationReport> {
 		selector: string,
 		pagination: PaginationParams<IModerationReport>,
 		options?: FindOptions<IModerationReport>,
-	): FindPaginated<FindCursor<Pick<IModerationReport, '_id' | 'message' | 'ts' | 'room'>>>;
+	): FindPaginated<FindCursor<Pick<MessageReport, '_id' | 'message' | 'ts' | 'room'>>>;
 
-	hideReportsByMessageId(messageId: IMessage['_id'], userId: string, reason: string, action: string): Promise<UpdateResult | Document>;
+	hideMessageReportsByMessageId(
+		messageId: IMessage['_id'],
+		userId: string,
+		reason: string,
+		action: string,
+	): Promise<UpdateResult | Document>;
 
-	hideReportsByUserId(userId: string, moderatorId: string, reason: string, action: string): Promise<UpdateResult | Document>;
+	hideMessageReportsByUserId(userId: string, moderatorId: string, reason: string, action: string): Promise<UpdateResult | Document>;
 }