+++ ee/apps/ddp-streamer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rocket.chat/ddp-streamer
 
+## 0.0.5
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.2.10
+- @rocket.chat/rest-typings@6.2.10
+- @rocket.chat/core-services@0.0.5
+- @rocket.chat/model-typings@0.0.5
+- @rocket.chat/models@0.0.5
+- @rocket.chat/instance-status@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes